wasmx-tendermint-p2p

+++ packages/wasmx-tendermint-p2p/assembly/types.ts
@@ -8,10 +8,10 @@ import { NodeInfo } from "wasmx-raft/assembly/types_raft";
 export class LogEntry {
     // this is also the block height
     index: i64;
-    termId: i32;
+    termId: i64;
     leaderId: i32;
     data: Base64String; // empty for finalized blocks;
-    constructor(index: i64, termId: i32, leaderId: i32, data: string) {
+    constructor(index: i64, termId: i64, leaderId: i32, data: string) {
         this.index = index;
         this.termId = termId;
         this.leaderId = leaderId;
@@ -24,10 +24,10 @@ export class LogEntry {
 export class LogEntryAggregate {
     // this is also the block height
     index: i64;
-    termId: i32;
+    termId: i64;
     leaderId: i32;
     data: wblocks.BlockEntry;
-    constructor(index: i64, termId: i32, leaderId: i32, data: wblocks.BlockEntry) {
+    constructor(index: i64, termId: i64, leaderId: i32, data: wblocks.BlockEntry) {
         this.index = index;
         this.termId = termId;
         this.leaderId = leaderId;
@@ -57,10 +57,10 @@ export class Transaction {
 // @ts-ignore
 @serializable
 export class TransactionResponse {
-    termId: i32;
+    termId: i64;
     leaderId: i32;
     index: i64;
-    constructor(termId: i32, leaderId: i32, index: i64) {
+    constructor(termId: i64, leaderId: i32, index: i64) {
         this.termId = termId;
         this.leaderId = leaderId;
         this.index = index;
@@ -71,12 +71,12 @@ export class TransactionResponse {
 @serializable
 export class AppendEntry {
     // leader’s term
-    termId: i32;
+    termId: i64;
     // so follower can redirect clients
     proposerId: i32;
     // block
     entries: LogEntryAggregate[]
-    constructor(termId: i32, proposerId: i32, entries: LogEntryAggregate[]) {
+    constructor(termId: i64, proposerId: i32, entries: LogEntryAggregate[]) {
         this.termId = termId;
         this.proposerId = proposerId;
         this.entries = entries;
@@ -87,11 +87,11 @@ export class AppendEntry {
 @serializable
 export class AppendEntryResponse {
     // currentTerm, for leader to update itself
-    termId: i32;
+    termId: i64;
     // true if follower contained entry matching prevLogIndex and prevLogTerm
     success: bool;
     lastIndex: i64;
-    constructor(termId: i32, success: bool, lastIndex: i64) {
+    constructor(termId: i64, success: bool, lastIndex: i64) {
         this.termId = termId;
         this.success = success;
         this.lastIndex = lastIndex;
